+++ src/Domains/Connectors/Intellicheck/Activities/IdVerificationReportActivity.php
@@ -5,14 +5,12 @@ declare(strict_types=1);
 namespace Kanvas\Connectors\Intellicheck\Activities;
 
 use Baka\Contracts\AppInterface;
-use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Notification;
 use Kanvas\Connectors\Intellicheck\Services\IdVerificationService;
+use Kanvas\Connectors\Intellicheck\Services\PeopleService;
 use Kanvas\Filesystem\Services\PdfService;
-use Kanvas\Guild\Customers\DataTransferObject\Address;
 use Kanvas\Guild\Customers\Models\People;
-use Kanvas\Locations\Models\Countries;
 use Kanvas\Notifications\Templates\Blank;
 use Kanvas\Users\Repositories\UsersRepository;
 use Kanvas\Workflow\Contracts\WorkflowActivityInterface;
@@ -82,7 +80,7 @@ class IdVerificationReportActivity extends KanvasActivity implements WorkflowAct
                     $people = $entity instanceof People ? $entity : $entity->people;
                     //update people name
                     if ($people instanceof People) {
-                        $this->updatePeopleInformation($people, $verificationData);
+                        PeopleService::updatePeopleInformation($people, $verificationData);
                     }
 
                     $usersToNotify = UsersRepository::findUsersByArray($entity->company->get('company_manager'), $app);
@@ -140,27 +138,5 @@ class IdVerificationReportActivity extends KanvasActivity implements WorkflowAct
                 'trace' => $e->getTraceAsString(),
             ];
         }
-    }
-
-    protected function updatePeopleInformation(
-        People $people,
-        array $verificationData
-    ): void {
-        $people->firstname = $verificationData['idcheck']['data']['firstName'] ?? $people->firstname;
-        $people->middlename = $verificationData['idcheck']['data']['middleName'] ?? $people->middlename;
-        $people->lastname = $verificationData['idcheck']['data']['lastName'] ?? $people->lastname;
-        $people->name = $verificationData['idcheck']['data']['firstName'] . ' ' . $verificationData['idcheck']['data']['lastName'];
-        $people->dob = isset($verificationData['idcheck']['data']['dateOfBirth'])
-            ? Carbon::createFromFormat('m/d/Y', $verificationData['idcheck']['data']['dateOfBirth'])->format('Y-m-d')
-            : $people->dob;
-        $people->saveOrFail();
-
-        $people->addAddress(new Address(
-            address: $verificationData['idcheck']['data']['address1'] ?? '',
-            city: $verificationData['idcheck']['data']['city'] ?? '',
-            state: $verificationData['idcheck']['data']['state'] ?? '',
-            country: $verificationData['idcheck']['data']['country'] ?? Countries::getByCode('US')->name,
-            zip: $verificationData['idcheck']['data']['postalCode'] ?? '',
-        ));
     }
 }
